refactor: pages를 views로 변경 (Next.js와 FSD 구조 통합)
- src/page-components 삭제 (기존 구조)
- src/views/ 생성 (flat 구조)
  - AboutPage.tsx
  - HomePage.tsx
  - PostsPage.tsx
  - PostDetailPage.tsx
  - CategoryPage.tsx
  - NotFoundPage.tsx
- app/ 폴더의 모든 import 경로 수정 (@/views)

Next.js Pages Router와 충돌 해결:
- src/pages는 Next.js가 자동 라우팅으로 인식
- src/views로 변경하여 FSD의 pages 레이어 역할 수행
