--- src/page-components/about/index.ts
@@ -1,1 +1,0 @@
-export { AboutPage } from './ui/AboutPage';
