--- src/page-components/category/index.ts
@@ -1,1 +1,0 @@
-export { CategoryPage } from './ui/CategoryPage';
