--- src/page-components/home/index.ts
@@ -1,1 +1,0 @@
-export { HomePage } from './ui/HomePage';
