--- src/page-components/index.ts
@@ -1,5 +1,0 @@
-export { HomePage } from './home';
-export { AboutPage } from './about';
-export { PostDetailPage } from './post-detail';
-export { CategoryPage } from './category';
-export { NotFoundPage } from './not-found';
