--- src/page-components/not-found/index.ts
@@ -1,1 +1,0 @@
-export { NotFoundPage } from './ui/NotFoundPage';
