--- src/page-components/post-detail/index.ts
@@ -1,1 +1,0 @@
-export { PostDetailPage } from './ui/PostDetailPage';
