--- src/page-components/posts/index.ts
@@ -1,1 +1,0 @@
-export { PostsPage } from './ui/PostsPage';
