+++ src/views/index.ts
@@ -1,0 +1,6 @@
+export { AboutPage } from './AboutPage';
+export { HomePage } from './HomePage';
+export { PostsPage } from './PostsPage';
+export { PostDetailPage } from './PostDetailPage';
+export { CategoryPage } from './CategoryPage';
+export { NotFoundPage } from './NotFoundPage';
